+++ R/GIRFandGFEVD.R
@@ -101,7 +101,7 @@
 #'   the transition weights across the Monte Carlo repetitions.
 #'
 #'   If \code{use_data_shocks=TRUE}, the GIRF is estimated using all, or a subset of, the length p histories in the data
-#'   as the initial values,  and using the sign and size of the corresponding structural shock recovered from the fitted model.
+#'   as the initial values, and using the sign and size of the corresponding structural shock recovered from the fitted model.
 #'   The subset of the length p histories are determined based in the settings given in the argument \code{data_girf_pars}.
 #'   Note that the arguments \code{shock_size}  and \code{init_regime} are ignored if \code{use_data_shocks=TRUE}.
 #' @return Returns a class \code{'girf'} list with the GIRFs in the first
@@ -469,10 +469,9 @@ GIRF <- function(stvar, which_shocks, shock_size=1, N=30, R1=250, R2=250, init_r
 #'        specified by the argument \code{init_regimes}. The number of initial values is set with the argument \code{R2}.}
 #'     \item{\code{"fixed"}:}{Estimate the GIRF for the initial values specified with the argument \code{init_values}.}
 #'   }
-#' @param use_data_shocks set \code{TRUE} for a special feature in which for every possible length \eqn{p} history in the data,
-#'   or a subset of them if so specified in the argument \code{data_gfevd_pars}, the GFEVD is estimated for a shock that
-#'   has the sign and size of the corresponding structural shock recovered from the data.
-#'   See the details section.
+#' @param use_data_shocks set \code{TRUE} (\textbf{recommended}) for a special feature in which for every possible length \eqn{p} history
+#'   in the data, or a subset of them if so specified in the argument \code{data_gfevd_pars}, the GFEVD is estimated for a shock that
+#'   has the sign and size of the corresponding structural shock recovered from the data. See the details section.
 #' @param data_gfevd_pars a length two numeric vector with the following elements determining settings for \code{initval_type="data"}
 #'   and \code{use_data_shocks=TRUE}:
 #'   \enumerate{
@@ -498,15 +497,14 @@ GIRF <- function(stvar, which_shocks, shock_size=1, N=30, R1=250, R2=250, init_r
 #'   corresponding structural shock recovered from the fitted model. This is done for every possible length \eqn{p} history
 #'   in the data, or to a subset of the histories based in the settings given in the argument \code{data_gfevd_pars}.
 #'   The GFEVD is then calculated as the average of the GFEVDs obtained from the GIRFs estimated for
-#'   the data shocks. The plot and print methods can be used as usual for this GFEVD. However, this feature also
-#'   obtain the contribution of each shock to the variance of the forecast errors at various horizons in specific
+#'   the data shocks. The plot and print methods can be used as usual for this GFEVD, but there is also a special feature
+#'   that allows to plot the contribution of each shock to the variance of the forecast errors at various horizons in specific
 #'   historical points of time. This can be done by using the plot method with the argument \code{data_shock_pars}.
-#'   Note that the arguments \code{shock_size}, \code{initval_type}, and \code{init_regime} are ignored if
-#'   \code{use_data_shocks=TRUE}.
+#'   Note that the arguments \code{shock_size}, \code{initval_type}, and \code{init_regime} are ignored if \code{use_data_shocks=TRUE}.
 #' @return Returns and object of class 'gfevd' containing the GFEVD for all the variables and to
 #'   the transition weights. Note that the decomposition does not exist at horizon zero for transition weights
 #'   because the related GIRFs are always zero at impact.
-#'   If \code{use_data_shocks=TRUE}, also contains the GFEVDs for each length \eqn{p} history in the data as
+#'   Also contains the individual GFEVDs for each used initial length \eqn{p} initial value (history) as
 #'   4D array with dimensions \code{[horizon, variable, shock, time]}.
 #' @seealso \code{\link{GIRF}}, \code{\link{linear_IRF}}, \code{\link{fitSSTVAR}}
 #' @references
@@ -573,9 +571,9 @@ GIRF <- function(stvar, which_shocks, shock_size=1, N=30, R1=250, R2=250, init_r
 #'  }
 #' @export
 
-GFEVD <- function(stvar, shock_size=1, N=30, initval_type=c("data", "random", "fixed"), use_data_shocks=FALSE,
-                  data_gfevd_pars=c(0, 0.75), R1=250, R2=250, init_regime=1, init_values=NULL, which_cumulative=numeric(0),
-                  ncores=2, burn_in=1000, exo_weights=NULL, seeds=NULL, use_parallel=TRUE) {
+GFEVD <- function(stvar, N=30, shock_size=1, initval_type=c("data", "random", "fixed"), which_cumulative=numeric(0), use_data_shocks=FALSE,
+                  data_gfevd_pars=c(0, 0.75), init_regime=1, init_values=NULL, R1=250, R2=250, ncores=2, burn_in=1000, exo_weights=NULL,
+                  seeds=NULL, use_parallel=TRUE) {
   check_stvar(stvar)
   initval_type <- match.arg(initval_type)
   cond_dist <- stvar$model$cond_dist
@@ -672,7 +670,6 @@ GFEVD <- function(stvar, shock_size=1, N=30, initval_type=c("data", "random", "f
     simulate.stvar(stvar, nsim=N+1, init_values=init_values_for_1girf, ntimes=R1, seed=seed, init_regime=init_regime,
                    burn_in=burn_in, exo_weights=exo_weights, girf_pars=list(shock_numb=shock_numb, shock_size=shock_size))
   }
-
   GIRF_shocks <- vector("list", length=d) # Storage for the GIRFs [[shock]]
 
   if(use_parallel) {
@@ -721,56 +718,54 @@ GFEVD <- function(stvar, shock_size=1, N=30, initval_type=c("data", "random", "f
   varnames <- c(varnames, paste0("tw", 1:M))
   shocknames <- paste0("Shock", 1:d)
 
-  GIRF_square_cumsum <- array(dim=c(N + 1, length(varnames), d),
-                              dimnames=list(0:N, varnames, shocknames)) # [horizon, variable, shock]
-  for(i1 in 1:d) { # Go through the shocks
-    res_in_array <- array(unlist(GIRF_shocks[[i1]]), dim=c(N + 1, length(varnames), R2))
-    if(length(which_cumulative) > 0) {
-      for(i2 in which_cumulative) {
-        res_in_array[, i2, ] <- apply(res_in_array[, i2, , drop=FALSE], MARGIN=3, FUN=cumsum) # Replace GIRF with cumulative GIRF
-      }
-    }
-    GIRF_square_cumsum[, , i1] <-  apply(apply(X=res_in_array, MARGIN=1:2, FUN=mean)^2, MARGIN=2, FUN=cumsum)
-  }
+  # GIRF_square_cumsum <- array(dim=c(N + 1, length(varnames), d),
+  #                             dimnames=list(0:N, varnames, shocknames)) # [horizon, variable, shock]
+  # for(i1 in 1:d) { # Go through the shocks
+  #   res_in_array <- array(unlist(GIRF_shocks[[i1]]), dim=c(N + 1, length(varnames), R2))
+  #   if(length(which_cumulative) > 0) {
+  #     for(i2 in which_cumulative) {
+  #       res_in_array[, i2, ] <- apply(res_in_array[, i2, , drop=FALSE], MARGIN=3, FUN=cumsum) # Replace GIRF with cumulative GIRF
+  #     }
+  #   }
+  #   GIRF_square_cumsum[, , i1] <-  apply(apply(X=res_in_array, MARGIN=1:2, FUN=mean)^2, MARGIN=2, FUN=cumsum)
+  # }
+  #
+  # GFEVD_results <- array(dim=c(N + 1, d, length(varnames)),
+  #                        dimnames=list(0:N, shocknames, varnames)) # [horizon, shock, variable]
+  # for(i1 in 1:ncol(GIRF_square_cumsum)) { # Go through the variables
+  #   denominator <- rowSums(GIRF_square_cumsum[, i1, ]) # The denominators for h=0,1,...,N
+  #   for(i2 in 1:d) { # Go through the shocks
+  #     GFEVD_results[ , i2, i1] <- GIRF_square_cumsum[, i1, i2]/denominator
+  #   }
+  # }
 
-  GFEVD_results <- array(dim=c(N + 1, d, length(varnames)),
-                         dimnames=list(0:N, shocknames, varnames)) # [horizon, shock, variable]
-  for(i1 in 1:ncol(GIRF_square_cumsum)) { # Go through the variables
-    denominator <- rowSums(GIRF_square_cumsum[, i1, ]) # The denominators for h=0,1,...,N
+  # Calculate the GFEVD also separately for each length p history a GIRF was estimated to
+  ind_GFEVD_results <- array(dim=c(N + 1, length(varnames), d, R2),
+                             dimnames=list(0:N, varnames, shocknames, 1:R2)) # [horizon, variable, shock, time])
+  for(i1 in 1:R2) { # Go through the length p histories
+    GIRF_square_cumsum_i1 <- array(NA, dim=c(N+1, length(varnames), d)) # [horizon, variable, shock] for time period i1
     for(i2 in 1:d) { # Go through the shocks
-      GFEVD_results[ , i2, i1] <- GIRF_square_cumsum[, i1, i2]/denominator
+      res_in_matrix <- GIRF_shocks[[i2]][[i1]] # [horizon, variable] GIRF of shock i2 at the time i1
+      if(length(which_cumulative > 0)) { # Any GIRFs that should be accumulated?
+        for(i3 in which_cumulative) {
+          res_in_matrix[, i3] <- cumsum(res_in_matrix[, i3]) # Replace GIRF with cumulative GIRF
+        }
+      }
+      # Cumulative squared GIRF of shock i2 at the time i1
+      GIRF_square_cumsum_i1[, , i2] <- apply(res_in_matrix^2, MARGIN=2, FUN=cumsum) # [horizon, variable]
+    }
+    # Calculate the denominators
+    for(i2 in 1:length(varnames)) { # Go through the variables
+      denominator <- rowSums(GIRF_square_cumsum_i1[, i2, ]) # The denominators for h=0,1,...,N (sums over socks)
+      for(i3 in 1:d) { # Go through the shocks
+        ind_GFEVD_results[ , i2, i3, i1] <- GIRF_square_cumsum_i1[, i2, i3]/denominator
+      }
     }
   }
 
-  # Is data shocks are used, calculate the GFEVD also separately for each length p history in the data
-  if(use_data_shocks) {
-    # Note that the dim is different to the other GFEVD_results below, because the interest is in the contribution of
-    # each shock to the variance of the forecast errors at various horizons in specific historical points of time.
-    data_GFEVD_results <- array(dim=c(N + 1, length(varnames), d, R2),
-                                dimnames=list(0:N, varnames, shocknames, 1:R2)) # [horizon, variable, shock, time])
-    for(i1 in 1:R2) { # Go through the length p histories
-      GIRF_square_cumsum_i1 <- array(NA, dim=c(N+1, length(varnames), d)) # [horizon, variable, shock] for time period i1
-      for(i2 in 1:d) { # Go through the shocks
-        res_in_matrix <- GIRF_shocks[[i2]][[i1]] # [horizon, variable] GIRF of shock i2 at the time i1
-        if(length(which_cumulative > 0)) { # Any GIRFs that should be accumulated?
-          for(i3 in which_cumulative) {
-            res_in_matrix[, i3] <- cumsum(res_in_matrix[, i3]) # Replace GIRF with cumulative GIRF
-          }
-        }
-        # Squared cumulative GIRF of shock i2 at the time i1
-        GIRF_square_cumsum_i1[, , i2] <- apply(res_in_matrix^2, MARGIN=2, FUN=cumsum) # [horizon, variable]
-      }
-      # Calculate the denonimators
-      for(i2 in 1:length(varnames)) { # Go through the variables
-        denominator <- rowSums(GIRF_square_cumsum_i1[, i2, ]) # The denominators for h=0,1,...,N (sums over socks)
-        for(i3 in 1:d) { # Go through the shocks
-          data_GFEVD_results[ , i2, i3, i1] <- GIRF_square_cumsum_i1[, i2, i3]/denominator
-        }
-      }
-    }
-  } else {
-    data_GFEVD_results <- NULL
-  }
+  # Calculate the GFEVD by calculating the mean over GFEVDs calculated for each length p history
+  GFEVD_results <- apply(ind_GFEVD_results, MARGIN=1:3, FUN=mean) # [horizon, variable, shock]
+  GFEVD_results <- aperm(GFEVD_results, perm=c(1, 3, 2)) # [horizon, shock, variable]
 
   if(use_parallel) message("Finished!")
   structure(list(gfevd_res=GFEVD_results,
@@ -784,7 +779,7 @@ GFEVD <- function(stvar, shock_size=1, N=30, initval_type=c("data", "random", "f
                  init_values=init_values,
                  seeds=seeds,
                  burn_in=burn_in,
-                 data_gfevd_res=data_GFEVD_results,
+                 ind_gfevd_res=ind_GFEVD_results,
                  stvar=stvar),
             class="gfevd")
 }